Moved HttpHeader into requests
